Sankey Diagram and Line Grapph
In this work, the production line of a hair cosmetics factory was studied. 
The factory operates from 8:00 AM to 5:15 PM, performing intervals from 9:00 AM to 9:20 AM, 11:30 AM to 12:35 PM, and 3:00 PM to 3:20 PM. Thus, based on this work period, simulations were performed for the case of three daily productions in order to comply with the demands presented
The study focuses on three daily productions, indicating that the factory aims to produce specific quantities of hair cosmetics products to meet demand. Each production cycle may involve different product types, batch sizes, or production schedules.